Add files via upload
Here are the new files, the New css and Js are for playing around to see if I can make something more out of this project. I am still working on the light/dark mode. That will be out no later than this week. I am also thinking of redesigning the products create page and  for product. that will be out till the end of next week.